+++ test/test_mixins/test_generic_commands.py
@@ -116,6 +116,17 @@ def test_sort_foo(r: redis.Redis):
         r.sort("foo", alpha=False)
 
 
+@pytest.mark.min_server("7")
+def test_sort_ro(r: redis.Redis):
+    r["score:1"] = 8
+    r["score:2"] = 3
+    r["score:3"] = 5
+    r.rpush("a", "3", "2", "1")
+    assert r.sort_ro("a", by="score:*") == [b"2", b"3", b"1"]
+    r.rpush("b", "2", "3", "1")
+    assert r.sort_ro("b", desc=True) == [b"3", b"2", b"1"]
+
+
 def test_sort_empty(r: redis.Redis):
     assert r.sort("foo") == []