+++ docs/about/changelog.md
@@ -9,6 +9,8 @@ description: Change log of all fakeredis releases
 
 - Implement support for hash expiration related commands @j00bar #328
     - `HEXPIRE`, `HEXPIREAT`, `HEXPIRETIME`, `HPERSIST`, `HPEXPIRE`, `HPEXPIREAT`, `HPEXPIRETIME`, `HPTTL`, `HTTL`,
+- Implement support for `SORT_RO` #325
+- 
       
 ## v2.24.1